model_training/out/requirements.txt
numpy
opencv-python
pillow
scikit-image
scipy
torch>=2.2
torchvision>=0.17
tqdm
albumentations